init dnd panel from backend
