+++ pom.xml
@@ -254,7 +254,11 @@
                 <artifactId>scaleph-plugin-seatunnel-native-flink</artifactId>
                 <version>${project.version}</version>
             </dependency>
-
+            <dependency>
+                <groupId>${project.groupId}</groupId>
+                <artifactId>scaleph-plugin-seatunnel-connectors</artifactId>
+                <version>${project.version}</version>
+            </dependency>
             <dependency>
                 <groupId>${project.groupId}</groupId>
                 <artifactId>scaleph-core-di</artifactId>
