+++ scaleph-api/src/main/java/cn/sliew/scaleph/api/controller/datadev/JobStepController.java
@@ -23,6 +23,7 @@ import cn.sliew.scaleph.common.enums.JobStepTypeEnum;
 import cn.sliew.scaleph.engine.seatunnel.service.SeatunnelConnectorService;
 import cn.sliew.scaleph.plugin.framework.core.PluginInfo;
 import cn.sliew.scaleph.plugin.framework.property.PropertyDescriptor;
+import cn.sliew.scaleph.system.dict.seatunnel.SeaTunnelPluginType;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -62,7 +63,7 @@ public class JobStepController {
     @ApiOperation(value = "查询 SeaTunnel connector 列表", notes = "查询 source, transform, sink 类型的 connector")
     public Set<PluginInfo> getAvailableConnectors(@ApiParam(name = "stepType",
             value = "job step 类型。source, trans, sink",
-            example = "source", allowableValues = "source, trans, sink") @PathVariable("stepType") JobStepTypeEnum stepType) {
+            example = "source", allowableValues = "source, trans, sink") @PathVariable("stepType") SeaTunnelPluginType stepType) {
         return seatunnelConnectorService.getAvailableConnectors(stepType);
     }
