+++ scaleph-engine/scaleph-engine-seatunnel/pom.xml
@@ -47,14 +47,20 @@
             <groupId>${project.parent.groupId}</groupId>
             <artifactId>scaleph-core-di</artifactId>
         </dependency>
+
         <dependency>
             <groupId>${project.parent.groupId}</groupId>
             <artifactId>scaleph-core-scheduler</artifactId>
         </dependency>
 
+<!--        <dependency>-->
+<!--            <groupId>${project.parent.groupId}</groupId>-->
+<!--            <artifactId>scaleph-plugin-seatunnel-native-flink</artifactId>-->
+<!--        </dependency>-->
+
         <dependency>
             <groupId>${project.parent.groupId}</groupId>
-            <artifactId>scaleph-plugin-seatunnel-native-flink</artifactId>
+            <artifactId>scaleph-plugin-seatunnel-connectors</artifactId>
         </dependency>
 
         <dependency>
