+++ scaleph-plugins/scaleph-plugin-seatunnel-connectors/src/main/java/cn/sliew/scaleph/plugin/seatunnel/flink/SeaTunnelConnectorPlugin.java
@@ -59,4 +59,5 @@ public abstract class SeaTunnelConnectorPlugin extends AbstractPlugin {
     }
 
     protected abstract SeaTunnelPluginMapping getPluginMapping();
+
 }
