+++ scaleph-plugins/scaleph-plugin-seatunnel-connectors/src/main/java/cn/sliew/scaleph/plugin/seatunnel/flink/connectors/assertion/sink/AssertProperties.java
@@ -16,29 +16,20 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.engine.seatunnel.service;
+package cn.sliew.scaleph.plugin.seatunnel.flink.connectors.assertion.sink;
 
-import cn.sliew.scaleph.common.enums.JobStepTypeEnum;
-import cn.sliew.scaleph.plugin.framework.core.PluginInfo;
-import cn.sliew.scaleph.plugin.framework.property.PropertyDescriptor;
-import cn.sliew.scaleph.plugin.seatunnel.flink.SeaTunnelConnectorPlugin;
-import cn.sliew.scaleph.system.dict.seatunnel.SeaTunnelPluginType;
+import cn.sliew.scaleph.plugin.framework.property.*;
 
-import java.util.List;
-import java.util.Properties;
-import java.util.Set;
+public enum AssertProperties {
+    ;
 
-/**
- * todo connector crud
- */
-public interface SeatunnelConnectorService {
-
-    List<PropertyDescriptor> getSupportedEnvProperties();
-
-    Set<PluginInfo> getAvailableConnectors(SeaTunnelPluginType stepType);
-
-    List<PropertyDescriptor> getSupportedProperties(String name);
-
-    SeaTunnelConnectorPlugin newConnector(String name, Properties properties);
+    public static final PropertyDescriptor<String> RULES = new PropertyDescriptor.Builder<String>()
+            .name("rules")
+            .description("Rule definition of user's available data. Each rule represents one field validation.")
+            .type(PropertyType.STRING)
+            .parser(Parsers.STRING_PARSER)
+            .properties(Property.Required)
+            .addValidator(Validators.NON_BLANK_VALIDATOR)
+            .validateAndBuild();
 
 }
