+++ scaleph-plugins/scaleph-plugin-seatunnel-connectors/src/main/java/cn/sliew/scaleph/plugin/seatunnel/flink/connectors/jdbc/sink/JdbcSinkPlugin.java
@@ -1,0 +1,22 @@
+package cn.sliew.scaleph.plugin.seatunnel.flink.connectors.jdbc.sink;
+
+import cn.sliew.scaleph.plugin.framework.core.PluginInfo;
+import cn.sliew.scaleph.plugin.seatunnel.flink.SeaTunnelConnectorPlugin;
+import cn.sliew.scaleph.plugin.seatunnel.flink.SeaTunnelPluginMapping;
+import com.google.auto.service.AutoService;
+
+@AutoService(SeaTunnelConnectorPlugin.class)
+public class JdbcSinkPlugin extends SeaTunnelConnectorPlugin {
+
+    public JdbcSinkPlugin() {
+        this.pluginInfo = new PluginInfo(getPluginName().getValue(),
+                "Jdbc Source Plugin , input records from jdbc connection.",
+                JdbcSinkPlugin.class.getName());
+    }
+
+    @Override
+    protected SeaTunnelPluginMapping getPluginMapping() {
+        return SeaTunnelPluginMapping.SINK_JDBC;
+    }
+
+}
