+++ scaleph-plugins/scaleph-plugin-seatunnel-connectors/src/main/java/cn/sliew/scaleph/plugin/seatunnel/flink/connectors/socket/source/SocketProperties.java
@@ -1,0 +1,43 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cn.sliew.scaleph.plugin.seatunnel.flink.connectors.socket.source;
+
+import cn.sliew.scaleph.plugin.framework.property.*;
+
+public enum SocketProperties {
+    ;
+
+    public static final PropertyDescriptor<String> HOST = new PropertyDescriptor.Builder<String>()
+            .name("host")
+            .description("socket server host")
+            .type(PropertyType.STRING)
+            .defaultValue("localhost")
+            .parser(Parsers.STRING_ARRAY_PARSER)
+            .addValidator(Validators.NON_BLANK_VALIDATOR)
+            .validateAndBuild();
+
+    public static final PropertyDescriptor<Integer> PORT = new PropertyDescriptor.Builder<Integer>()
+            .name("port")
+            .description("socket server port")
+            .type(PropertyType.INT)
+            .defaultValue(9999)
+            .parser(Parsers.INTEGER_PARSER)
+            .addValidator(Validators.NON_NEGATIVE_INTEGER_VALIDATOR)
+            .validateAndBuild();
+}
