+++ scaleph-support/scaleph-system/src/main/java/cn/sliew/scaleph/system/dict/seatunnel/SeaTunnelPluginType.java
@@ -28,7 +28,7 @@ public enum SeaTunnelPluginType implements DictInstance {
 
     SOURCE("source", "Source"),
     SINK("sink", "Sink"),
-
+    TRANSFORM("trans", "Trans"),
     ;
 
     @JsonValue
